requirements.txt
spotipy==2.23.0
python-dotenv==1.0.0
flask==2.3.3
werkzeug==2.3.7